+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/AuthTests.java
@@ -76,7 +76,7 @@ public class AuthTests extends BaseIntegrationTest {
                             }
 
                             @Override
-                            public void onClientResetRequired(SyncSession session, ClientResetHandler error) {
+                            public void onClientResetRequired(SyncSession session, ClientResetHandler handler) {
                                 fail("Client Reset");
                             }
                         })