+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/ManagementRealmTests.java
@@ -69,7 +69,7 @@ public class ManagementRealmTests extends BaseIntegrationTest {
                     }
 
                     @Override
-                    public void onClientResetRequired(SyncSession session, ClientResetHandler error) {
+                    public void onClientResetRequired(SyncSession session, ClientResetHandler handler) {
                         fail("Client Reset");
                     }
                 })